moved utility functions around
